+++ Source/CursesDialog/form/frm_req_name.c
@@ -153,7 +153,7 @@ int form_request_by_name( const char *str )
       strncpy(buf,str,sizeof(buf));
       while( (i<sizeof(buf)) && (buf[i] != '\0') )
 	{
-	  buf[i] = toupper(buf[i]);
+	  buf[i] = toupper((int)(buf[i]));
 	  i++;
 	}
